+++ Changes.md
@@ -1,3 +1,4 @@
+* #461 - use tournament information when looking up playoff brackets to ensure score sheets have correct bracket numbers
 * Take footer off of the slide show page
 
 Release 10.5